requirements.txt
requests
pyobjc
beautifulsoup4